Improved introduction and core concepts (#1116). Also added installation instructions for different platforms.
