+++ documentation/installing/installing-linux.asciidoc
@@ -1,0 +1,119 @@
+---
+title: Installing Tools in Linux
+order: 50
+layout: page
+---
+
+[[installing.linux]]
+= Installing Tools in Linux
+
+This section contains instructions for some Debian- and RPM-based distributions.
+The installation should be very similar on other Linux distributions.
+
+== Debian and Ubuntu
+
+On Debian and Ubuntu, it is easiest to use OpenJDK, as it is available in the standard  package repositories.
+Similarly, Maven, Git, and Node.js are available from a repository.
+The Node.js package includes npm, the Node package manager used for handling frontend dependencies.
+
+. Install Node.js package sources
++
+Most distributions do not have a sufficiently new Node.js version through their package manager.
+Begin by adding the latest Node.js LTS repository to your APT sources:
++
+----
+$ curl -sL https://deb.nodesource.com/setup_12.x | sudo -E bash -
+----
+
+. Install the development tools
++
+You can install all the needed development tools easily with:
++
+----
+$ sudo apt-get install -y openjdk-11-jdk maven git nodejs
+----
+
+If your system does not have `sudo` installed, execute the above commands as root.
+
+== Fedora and RHEL
+
+On Fedora (since 18) and RHEL (since 8), you use the DNF package manager.
+
+. Install the development tools:
++
+----
+$ sudo dnf -y install java-11-openjdk.x86_64 maven nodejs
+----
+
+. Ensure that you are running Java 11 by running:
++
+----
+$ java -version
+----
++
+If you are not using 11, change it with:
++
+----
+$ sudo alternatives --config java
+----
++
+Select OpenJDK 11.
+
+== RPM-based Systems Using YUM
+
+. Install Node.js package sources
++
+Most distributions do not have a sufficiently new Node.js version through their package manager.
+Begin by adding the latest Node.js LTS repository to your APT sources:
++
+----
+$ curl -sL https://deb.nodesource.com/setup_12.x | sudo -E bash -
+----
+
+. Install the development tools:
++
+----
+$ sudo yum install -y java-11-openjdk-devel.x86_64 maven git nodejs
+----
+
+. Ensure that you are running Java 11 by running:
++
+----
+$ java -version
+----
++
+If you are not using 11, change it with:
++
+----
+$ sudo alternatives --config java
+----
+
+== OpenSuSE
+
+On OpenSuSE, you use the zypper package manager as follows:
+
+----
+$ sudo zypper install java-11-openjdk maven nodejs12
+----
+
+Depending on your system version, you may need to install package repositories or do some other steps.
+
+== Arch Linux
+
+On Arch Linux, you use the pacman package manager as follows:
+
+----
+$ pacman -S jdk11-openjdk maven nodejs
+----
+
+Depending on your system version, you may need to install package repositories or do some other steps.
+
+== Installing Amazon Corretto
+
+If you want to install Amazon Corretto rather than OpenJDK, you either have to add the package repository or install it from a DEB, RPM, or a tar package.
+
+See:
+
+* Instructions for installing from APT (Debian) or YUM (RPM) repository in https://docs.aws.amazon.com/corretto/latest/corretto-11-ug/linux-info.html[Amazon Corretto 11 Guide for Linux]
+
+* Download installation packages from link:https://docs.aws.amazon.com/corretto/latest/corretto-11-ug/downloads-list.html[https://docs.aws.amazon.com/corretto/latest/corretto-11-ug/downloads-list.html]
