+++ documentation/installing/installing-macos.asciidoc
@@ -1,0 +1,76 @@
+---
+title: Installing Tools in macOS
+order: 40
+layout: page
+---
+
+[[installing.macos]]
+= Installing Tools in macOS
+:experimental:
+
+In the following, we guide you through installing the development tools in the macOS operating system.
+
+[[installing.macos.java]]
+== Installing Java
+
+To install Java in macOS:
+
+. Go to the Amazon Corretto 11 download page at https://aws.amazon.com/corretto/[aws.amazon.com/corretto].
+
+. Download and run the macOs installer (.pkg).
+
+. Follow the prompts in the wizard.
++
+image:images/mac-corretto-installer.png[Corretto Installer]
+
+[[installing.macos.homebrew]]
+== Installing Homebrew
+
+Homebrew is a package manager that gives the easiest way to install both Maven and Node.js on macOS.
+
+To install Homebrew, copy and paste the following into a terminal window:
+
+----
+$ /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+----
+
+[[installing.macos.maven]]
+== Installing Maven and Node.js
+
+To install both Maven and Node.js, enter the following in a terminal window:
+
+----
+$ brew install node maven
+----
+
+== Installing the Eclipse IDE (optional)
+
+You can install the Eclipse IDE either using an installer or manually.
+The installer downloads the installation package, extracts it, and sets it up in the system.
+You can also choose to manually download and extract the software package and then set it up manually.
+
+After installing the Eclipse IDE itself, you can install the Vaadin Plugin for Eclipse.
+
+To install the Eclipse IDE using the installer and the the Vaadin plugin:
+
+. Download the Eclipse Installer package for Mac OS X from
+link:https://www.eclipse.org/downloads/packages/installer[eclipse.org/downloads/packages/installer].
+
+. Follow the instructions in the installer download page
+
+. Start Eclipse
+
+. Install and configure the Vaadin Plugin for the Eclipse IDE, as described in <<installing-eclipse#, "Installing Vaadin Plugin for the Eclipse IDE">>.
+
+== Installing IntelliJ IDEA (optional)
+
+. Open a browser and go to the IntelliJ IDEA Mac download page at https://www.jetbrains.com/idea/download/[jetbrains.com/idea/download].
+
+. Download the Community Edition macOS Disk Image file (.dmg).
+
+. Mount it as another disk in your system.
+
+. Copy IntelliJ IDEA to your Applications folder.
+
+. Start IntelliJ IDEA and set up your preferences.
+You can use the defaults, unless you have a reason not to.
