+++ documentation/installing/installing-windows.asciidoc
@@ -1,0 +1,108 @@
+---
+title: Installing Tools in Windows
+order: 30
+layout: page
+---
+
+[[installing.windows]]
+= Installing Tools in Windows
+:experimental:
+
+In the following, we guide you through installing the development tools in the Windows operating system.
+
+[[installing.windows.java]]
+== Installing Java
+
+To install Java in Windows:
+
+. Go to the Amazon Corretto 11 download page at https://aws.amazon.com/corretto/[aws.amazon.com/corretto].
+
+. Download and run the Windows installer (.msi).
+
+. Follow the prompts in the wizard.
++
+image:images/win-coretto-installer.png[Corretto Installer]
+
+[[installing.windows.maven]]
+== Installing Maven
+
+In the following, we describe how to install Maven and include it in the system path:
+
+. Open a browser and go to the Maven download page at https://maven.apache.org/download.cgi[maven.apache.org/download.cgi].
+
+. Download the *Binary zip archive* in the *Files* section.
+
+. Extract the archive to `C:\Program Files`.
+
+. Open Windows *Control Panel*
+
+. Select *menu:Systems and Security[System > Advanced system settings]*.
+
+. Click [guibutton]#Environment Variables#.
+
+. In *Environment Variables*, select the *Path* systems variable (in the bottom box) and click [guibutton]#Edit#.
+
+. Click [guibutton]#Browse# in the edit dialog.
+
+. Navigate to and select `C:\Program Files\apache-maven-3.6.3\bin` (substitute the version number you downloaded) and click [guibutton]#OK#.
+
+. Click [guibutton]#OK# in all open dialogs to close them and save the environment variable.
+
+[[installing.windows.node]]
+== Installing Node.js
+
+To install Node.js:
+
+. Open a browser and go to the Node.js download page at https://nodejs.org/en/download/[nodejs.org/en/download].
+
+. Download and run the *Windows Installer* (.msi) for your system.
+
+. Follow the prompts in the wizard.
+
+== Installing Git
+
+Git is the currently most used software revision control system, so you most likely will need it for your own projects as well.
+It is not needed for using Vaadin.
+You may find using Git to clone Vaadin starter repositories useful.
+
+To install Git:
+
+. Open a browser and go to the Git download page at https://git-scm.com/downloads[git-scm.com/downloads].
+
+. Download and run the Windows installer (.exe) for your system.
+
+. Follow the prompts in the wizard.
+If you are unsure about any option, use the defaults.
+
+== Installing the Eclipse IDE
+
+You can install the Eclipse IDE either using an installer or manually.
+The installer downloads the installation package, extracts it, and sets it up in the system.
+You can also choose to manually download and extract the software package and then set it up manually.
+
+After installing the Eclipse IDE itself, you can install the Vaadin Plugin for Eclipse.
+
+To install the Eclipse IDE using the installer and the the Vaadin plugin:
+
+. Download the Eclipse Installer package for Windows from
+link:https://www.eclipse.org/downloads/packages/installer[eclipse.org/downloads/packages/installer].
+
+. Follow the instructions in the installer download page
+
+. Start Eclipse
+
+. Install and configure the Vaadin Plugin for the Eclipse IDE, as described in <<installing-eclipse#, "Installing Vaadin Plugin for the Eclipse IDE">>.
+
+
+== Installing IntelliJ IDEA
+
+. Open a browser and go to the IntelliJ Windows download page at https://www.jetbrains.com/idea/download/[jetbrains.com/idea/download].
+
+. Download and run the Community Edition installer (.exe).
+
+. Follow the prompts in the wizard.
+
+. Reboot your computer to finish the setup.
+
+.  Start IntelliJ and set up your preferences.
+You can use the defaults, unless you have a reason not to.
