+++ documentation/src/main/html/IntroductionWhatIsVaadin.html
@@ -1,0 +1,6 @@
+tutorial::introduction/introduction-overview.asciidoc
+
+<div>
+  <vaadin-text-field></vaadin-text-field>
+  <vaadin-button>Say hello</vaadin-button>
+</div>
